# openapi.yaml
openapi: 3.0.3
info:
  title: Diary API
  version: 2.0.0

paths:
  /auth/signup:
    post:
      summary: アカウント作成
      description: メールアドレスとパスワードを使用して新しいアカウントを作成します。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '201':
          description: アカウント作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
        '400':
          description: バリデーションエラー（メール重複など）
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: サーバーエラー
  
  /auth/signin:
    post:
      summary: アカウント認証（ログイン）
      description: |
        入力されたメールアドレスとパスワードを users テーブルと照合し、
        一致した場合はログイン成功としてユーザーIDを返す。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: "Login successful"
        '401':
          description: メールアドレスまたはパスワードが不正
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid email or password"
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"



  /posts/today:
    get:
      summary: 今日の投稿を取得
      description: |
        サーバー側で日付ライブラリを使って「今日」の日付を決定し、その日付に紐づく投稿を返す。
        クライアントは日付を指定する必要はない。存在しなければ null を返す。
      responses:
        '200':
          description: 投稿データ
          content:
            application/json:
              schema:
                type: object
                nullable: true
                properties:
                  id:
                    type: integer
                    example: 1
                  date:
                    type: string
                    format: date
                    example: "2025-09-19"
                  text:
                    type: string
                    example: "今日は公園に行った。"
    post:
      summary: 今日の投稿を新規作成または更新
      description: |
        サーバー側で日付ライブラリを使って「今日」の日付を決定し、その日付をキーに投稿を新規作成または更新する。
        クライアントは日付を指定する必要はなく、本文(text)のみ送信すればよい。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: "今日は図書館に行った。"
      responses:
        '200':
          description: 投稿が保存された
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  date:
                    type: string
                    format: date
                  text:
                    type: string

  /posts/month:
    get:
      summary: 指定した月の投稿日付とID一覧を取得
      parameters:
        - in: query
          name: year
          schema:
            type: integer
          required: true
          example: 2025
        - in: query
          name: month
          schema:
            type: integer
          required: true
          example: 9
      responses:
        '200':
          description: 投稿日付とIDの配列
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    date:
                      type: string
                      format: date

  /posts/{id}:
    get:
      summary: 投稿詳細を取得
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: 投稿詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  date:
                    type: string
                    format: date
                  text:
                    type: string

    post:
      summary: 投稿要約を取得
      description: text を生成AIに投げて要約を返す。
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: 要約
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    example: "今日は公園に行ったことについての日記"